pyproject.toml
[project]
name = "f1-chatter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.8",
    "django-ninja>=1.1.0",
    "gunicorn>=23.0.0",
    "fastf1>=3.1.0",
]

[tool.ruff.lint]
select = ["E", "F", "W"]
